crawler: dump data for invalid PR and review
