+++ tests/unit/test_crawler_gerrit.py
@@ -34,7 +34,7 @@ class TestGerritCrawler(unittest.TestCase):
         input_review, xtrd_ref = load_change(name)
 
         rf = review.ReviewesFetcher(base_url, None)
-        xtrd = rf.extract_objects([input_review])
+        xtrd = rf.extract_objects([input_review], print)
 
         ddiff = DeepDiff(xtrd_ref, xtrd, ignore_order=True)
         if ddiff:
